+++ package.json
@@ -38,6 +38,8 @@
     "arlas-web-contributors": "7.4.*",
     "cli-color": "^1.2.0",
     "core-js": "^2.4.1",
+    "ng-pick-datetime": "^5.2.6",
+    "ng-pick-datetime-moment": "^1.0.7",
     "rxjs": "^5.5.6",
     "zone.js": "^0.8.19"
   },
