+++ src/app/components/timeline/date-picker/date-picker.component.css
@@ -1,0 +1,91 @@
+.arlas-time-shortcuts--container {
+}
+
+.arlas-date-picker-input{
+    background-color: #FF4081;
+    border: none;
+    color: white;
+    font-size: 0.9em;
+    padding-left: 5px;
+    padding-top: 8px;
+    cursor: pointer;
+    
+
+
+}
+
+input:focus {
+    outline: 0px;
+  }
+
+.shortcut-button {
+  background-color: #FF4081;
+  float: right;
+  display: flex;
+  height: 30px;
+  justify-content: center;
+  align-items: center;
+  border-top-left-radius: 2px;
+  border-top-right-radius: 2px;
+  padding: 0 3px 0 3px;
+}
+
+.shortcut-button__icon {
+  color: white;
+  width: 18px;
+  height: 18px;
+  font-size: 18px;
+  background-color: #FF4081;
+
+}
+
+.cursor {
+  cursor: pointer;
+}
+
+.shortcut-remove-filter {
+  color: white;
+  width: 15px;
+  height: 15px;
+  font-size: 15px;
+  font-weight: bold;
+  margin-left: 5px;
+  cursor: pointer;
+}
+
+.shortcut-button__label {
+  font-size: 1em;
+  color: white;
+  padding-left: 5px;
+  padding-top: 2px;
+  background-color: #FF4081;
+
+}
+
+.arlas-time-shortcuts__list--container {
+  float: right;
+  display: inline-flex;
+  padding-top: 5px;
+  padding-left: 5px;
+  font-size: 0.8em;
+  background-color: white;
+  box-shadow: 0 3px 1px -2px rgba(0,0,0,.2), 0 2px 2px 0 rgba(0,0,0,.14), 0 1px 5px 0 rgba(0,0,0,.12);
+}
+
+.arlas-time-shortcuts__list {
+  list-style-type: none;
+  margin-bottom: 0;
+  padding-right: 20px;
+  padding-left: 0;
+}
+
+.arlas-time-shortcut {
+  color: rgb(48, 136, 177);
+  cursor: pointer;
+}
+
+.arlas-time-shortcut:hover {
+  text-decoration: underline;
+}
+
+
