--- src/app/components/timeline/timeline-shortcut.component.html
@@ -1,14 +1,0 @@
-<div class="arlas-time-shortcuts--container">
-  <div>
-      <mat-basic-chip class="shortcut-button">
-          <mat-icon class="shortcut-button__icon" [class.cursor]="timeShortcuts?.length > 0" (click)="showSortcuts()" matTooltip="{{HIDE_SHOW + ' time shortcuts' | translate}}" [matTooltipPosition]="'above'">access_time</mat-icon>
-          <span class="shortcut-button__label">{{timelineContributor?.timeLabel | getLabel: dateFormat:translate}}</span>
-          <mat-icon *ngIf="showRemoveIcon" (click)="removeTimelineCollaboration()" class="shortcut-remove-filter" matTooltip="{{'Remove time filter' | translate}}" [matTooltipPosition]="'before'" >clear</mat-icon>
-      </mat-basic-chip>
-  </div>
-  <div *ngIf="showShortcuts" class="arlas-time-shortcuts__list--container">
-    <ul class="arlas-time-shortcuts__list" *ngFor="let shortcutType of getKeys(timeShortcutsMap)">
-      <li class="arlas-time-shortcut" *ngFor="let shortCut of timeShortcutsMap.get(shortcutType)" (click)="setShortcut(shortCut)">{{shortCut.label}}</li>
-    </ul>
-  </div>
-</div>
