+++ src/app/components/timeline/timeline-shortcut/timeline-shortcut.component.css
