+++ src/app/components/timeline/timeline/timeline.component.css
