+++ src/app/components/timeline/timeline/timeline.component.spec.ts
@@ -18,15 +18,22 @@
  */
 
 import { ComponentFixture, TestBed, async } from '@angular/core/testing';
-import { MatButtonModule, MatCardModule, MatExpansionModule, MatIconModule, MatSelectModule,
-   MatTooltipModule, MatChipsModule } from '@angular/material';
+import {
+  MatButtonModule, MatCardModule, MatExpansionModule, MatIconModule, MatSelectModule,
+  MatTooltipModule, MatChipsModule
+} from '@angular/material';
 import { BrowserModule } from '@angular/platform-browser';
 import { DonutModule, HistogramModule, PowerbarsModule, ResultsModule } from 'arlas-web-components';
-import { ArlasCollaborativesearchService, ArlasStartupService, ArlasConfigService } from '../../services/startup/startup.service';
-import { WidgetComponent } from '../widget/widget.component';
-import { TimelineComponent, TimelineShortcutComponent, LabelPipe } from './timeline.component';
+import { ArlasCollaborativesearchService, ArlasStartupService, ArlasConfigService } from '../../../services/startup/startup.service';
+import { WidgetComponent } from '../../widget/widget.component';
+import { TimelineComponent } from './timeline.component';
 import { TranslateModule, TranslateService, TranslateLoader, TranslateFakeLoader } from '@ngx-translate/core';
 import { HttpModule } from '@angular/http';
+import { TimelineShortcutComponent } from '../../timeline/timeline-shortcut/timeline-shortcut.component';
+import { GetTimeLabelPipe } from '../../../pipes/get-time-label.pipe';
+import { DatePickerComponent } from 'app/components/timeline/date-picker/date-picker.component';
+import { OwlNativeDateTimeModule, OwlDateTimeModule } from 'ng-pick-datetime';
+import { FormsModule } from '@angular/forms';
 
 describe('TimelineComponent', () => {
   let component: TimelineComponent;
@@ -34,9 +41,10 @@ describe('TimelineComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ TimelineComponent, WidgetComponent, TimelineShortcutComponent, LabelPipe ],
+      declarations: [TimelineComponent, DatePickerComponent, WidgetComponent, TimelineShortcutComponent, GetTimeLabelPipe],
       imports: [
         MatCardModule, MatIconModule, MatExpansionModule, MatSelectModule, MatButtonModule, MatChipsModule,
+        OwlDateTimeModule, OwlNativeDateTimeModule, FormsModule,
         MatTooltipModule, BrowserModule, HistogramModule, ResultsModule, PowerbarsModule, DonutModule, HttpModule,
         TranslateModule.forRoot({
           loader: { provide: TranslateLoader, useClass: TranslateFakeLoader }
@@ -46,7 +54,7 @@ describe('TimelineComponent', () => {
         ArlasCollaborativesearchService, ArlasStartupService, ArlasConfigService, TranslateService
       ]
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {
