+++ src/app/pipes/get-time-label.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { GetTimeLabelPipe } from './get-time-label.pipe';
+
+describe('GetTimeLabelPipe', () => {
+  it('create an instance', () => {
+    const pipe = new GetTimeLabelPipe();
+    expect(pipe).toBeTruthy();
+  });
+});
