+++ src/assets/i18n/en.json
@@ -10,7 +10,7 @@
     "Archivage": "Archivage",
     "Reception": "Reception"
   },
-  "Swimlane" : {
+  "Swimlane": {
     "Mission": "Mission",
     "Archivage": "Archivage",
     "Reception": "Reception"
@@ -19,5 +19,27 @@
     "title": "Geoname"
   },
   "From": "From",
-  "to": "to"
+  "to": "to",
+  "Datepicker": {
+    "upSecondLabel": "Add a second",
+    "downSecondLabel": "Minus a second",
+    "upMinuteLabel": "Add a minute",
+    "downMinuteLabel": "Minus a minute",
+    "upHourLabel": "Add a hour",
+    "downHourLabel": "Minus a hour",
+    "prevMonthLabel": "Previous month",
+    "nextMonthLabel": "Next month",
+    "prevYearLabel": "Previous year",
+    "nextYearLabel": "Next year",
+    "prevMultiYearLabel": "Previous 21 years",
+    "nextMultiYearLabel": "Next 21 years",
+    "switchToMonthViewLabel": "Change to month view",
+    "switchToMultiYearViewLabel": "Choose month and year",
+    "cancelBtnLabel": "Cancel",
+    "setBtnLabel": "OK",
+    "rangeFromLabel": "From",
+    "rangeToLabel": "To",
+    "hour12AMLabel": "AM",
+    "hour12PMLabel": "PM"
+  }
 }
